everything corrected and souted for usefulness

+++ basics/basiclibrary/src/main/java/basiclibrary/Averages.java
@@ -1,7 +1,7 @@
 package basiclibrary;
 
 public class Averages {
-  public static double averageFinder(double[] averagePool) {
+  public static double averageFinder(int[] averagePool) {
     double total = 0;
     for (int i = 0; i < averagePool.length; i++) {
       total += averagePool[i];
